Set automatic draw order to avoid transparent objects occluding others
The 'ORDERING_ENABLED' had been set to false. Normally, this will auto-order
objects from back-to-front to ensure objects are drawn correctly during render
times. Setting this to false appears to render them in the order they are listed
in the object file, which can allow the user to view them from the 'wrong' angle,
causing the 'wrong' draw order and some objects to simply not show up.

The config option was set in the initial commit of this git project, and no changes
appear to have been made to the X Toolkit since then. That makes me think there were
other problems the original author was addressing.

One problem I find when ORDERING_ENABLED is True is objects 'popping' or occationally
'flickering' when the view is changed. I've only noticed this when two objects intersect
and the ordering algorithm is having trouble deciding which object is on top, similar to
the Z-Fighting problem.

I didn't find any documentation on the ORDERING_ENABLED config option, but I did find
the only commit in the X toolkit mentioning it:

https://github.com/xtk/X/commit/ad2eeb071c7ec64d0a71dc9715ef8844b616fdbd